refactor: extract Dump methods into format.go
parser.go shrinks from 778 to 167 lines by moving all formatter code
(every node's Dump method, NodeList.Dump, ConfiguredNodeList.Dump,
IndentConfig, the indentConfig global, isTemplateElement helper) into
a dedicated format.go.

The AST type definitions stay in parser.go alongside the public entry
points (NewParser, NewParserWithConfig, NewAdminParser, NewStore-
frontParser), isVoidElement, and TraverseNode. Methods remain attached
to their types — Go lets us split type and method across files in the
same package.

No behavior change: all 45 testdata fixtures continue to format byte-
identically and the shopware/storefront smoke test shows the same
0-regression result as before the move.

Dead-code cleanup: fromEntitiesToText (used only by the deleted legacy
parseAttrValue) is removed. htmlCommentStart was already gone.

SetIndentConfig is marked Deprecated in its doc comment, pointing
callers toward future explicit-Formatter use.
